lib: Fix memory corruption when releasing route-map hash
Previously, if a route-map is created and deleted multiple times within
an update timer interval, we rely on the order of elements in hash
table to get the correct one freed. However, if hash size threshold is
hit, the list is reversed in hash_expand(), which will cause memory
corruption. This commit ensures the exact route-map element is released
from hash table.

Signed-off-by: Xiao Liang <shaw.leon@gmail.com>